Extract Guided Tutorial root
